+++ src/main/java/_persistence/rubric/RubricQueryBuilder.java
@@ -20,9 +20,8 @@ class RubricQueryBuilder extends PreparedQueryBuilder {
   /**
    * The [buildRetrieveRubricsByCourseIdQuery] method...
    */
-  PreparedStatement buildRetrieveRubricsByCourseIdQuery (
-    String courseId
-  ) throws SQLException {
+  PreparedStatement buildRetrieveRubricsByCourseIdQuery (String courseId)
+      throws SQLException {
     String query = "SELECT * FROM rubric WHERE course_pk1 = ?";
 
     PreparedStatement statement = connection.prepareStatement (query);
@@ -34,9 +33,8 @@ class RubricQueryBuilder extends PreparedQueryBuilder {
   /**
    * The [buildRetrieveRubricByIdQuery] method...
    */
-  PreparedStatement buildRetrieveRubricByIdQuery (
-    String rubricId
-  ) throws SQLException {
+  PreparedStatement buildRetrieveRubricByIdQuery (String rubricId)
+      throws SQLException {
     String query = "SELECT * FROM rubric WHERE pk1 = ?";
 
     PreparedStatement statement = connection.prepareStatement (query);
@@ -48,11 +46,24 @@ class RubricQueryBuilder extends PreparedQueryBuilder {
   /**
    * The [buildRetrieveColumnsByRubricIdQuery] method...
    */
-  PreparedStatement buildRetrieveColumnsByRubricIdQuery (
-    String rubricId
-  ) throws SQLException {
+  PreparedStatement buildRetrieveColumnsByRubricIdQuery (String rubricId)
+      throws SQLException {
     String query =
       "SELECT * FROM rubric_column WHERE rubric_pk1 = ? GROUP BY rubric_pk1";
+
+    PreparedStatement statement = connection.prepareStatement (query);
+    statement.setString (1, rubricId.split ("_")[1]);
+
+    return statement;
+  }
+
+  /**
+   * The [buildRetrieveRowsByRubricIdQuery] method...
+   */
+  PreparedStatement buildRetrieveRowsByRubricIdQuery (String rubricId)
+      throws SQLException {
+    String query =
+      "SELECT * FROM rubric_row WHERE rubric_pk1 = ? GROUP BY rubric_pk1";
 
     PreparedStatement statement = connection.prepareStatement (query);
     statement.setString (1, rubricId.split ("_")[1]);